requirements.txt
python-dotenv==1.0.0           # For loading environment variables from a .env file
boto3==1.26.5                  # For AWS SDK (WorkMail integration)
requests==2.31.0               # For making HTTP requests (used for Gmail, Outlook, etc. integrations)
pydantic==2.0.3                # For validating and parsing environment variables (optional but recommended)
bleach==5.0.0                  # Check for the latest version
google-auth==2.31.0
google-auth-oauthlib==0.5.1
google-auth-httplib2==0.1.0
google-api-python-client==2.93.0